requirements.txt
# 零差錯 AI 語音點餐系統 - Python 依賴包

# Web 框架
Flask==2.3.3
Flask-CORS==4.0.0

# 環境變量管理
python-dotenv==1.0.0

# Azure Speech Services SDK
azure-cognitiveservices-speech==1.34.0

# OpenAI 客戶端（用於 OpenRouter API）
openai

# HTTP 請求庫
requests==2.31.0

# 數據處理
dataclasses-json==0.6.1

# 日誌和調試
colorlog==6.7.0

# WebSocket 支援（將在後續任務中使用）
Flask-SocketIO==5.3.6
python-socketio==5.9.0

# 數據庫支援（將在後續任務中使用）
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# 音頻處理（將在後續任務中使用）
pydub==0.25.1

# 測試框架
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# 開發工具
black==23.11.0
flake8==6.1.0
isort==5.12.0

# 生產環境部署（可選）
gunicorn==21.2.0
gevent==23.9.1